+++ src/features/wallet/screens/TransferDetails/form.ts
@@ -7,15 +7,14 @@ export const getTransferFormInitialValues = (): TransferFormValues => ({
   amount: '',
 });
 
-export const useTransferFormValidation = (maxAmount: number) =>
+export const useTransferFormValidation = () =>
   useMemo(
     () =>
       object().shape({
         amount: number()
           .typeError('Amount must be a number')
           .notOneOf([0], 'Amount cannot be 0')
-          .max(maxAmount, `Amount must not exceed ${maxAmount}`)
           .required('Amount is required'),
       }),
-    [maxAmount],
+    [],
   );